Fix some watchdog problems

+++ tpu_runner.py
@@ -532,12 +532,11 @@ class TPURunner(object):
     def _watchdog(self):
         self.watchdog_time = time.time()
         while not self.watchdog_shutdown:
-            if self.pipe and self.pipe.first_name is None and \
+            if self.pipe and self.pipe.first_name is not None and \
                 time.time() - self.watchdog_time > self.max_idle_secs_before_recycle:
                 logging.warning("No work in {} seconds, watchdog shutting down TPUs.".format(self.max_idle_secs_before_recycle))
                 self.runner_lock.acquire(timeout=MAX_WAIT_TIME)
-                if self.pipe:
-                    self.pipe.delete()
+                self.pipe.delete()
                 self.runner_lock.release()
                 # Pipeline will reinitialize itself as needed
             time.sleep(self.watchdog_idle_secs)
